// src/data/enums/EventType.ts
export enum EventType {
    RESIZE = "resize",
    MOUSE_UP = "mouseup",
    MOUSE_DOWN = "mousedown",
    MOUSE_MOVE = "mousemove",
    MOUSE_WHEEL = "wheel",
    POINTER_DOWN = "pointerdown",
    POINTER_MOVE = "pointermove",
    POINTER_UP = "pointerup",
    TOUCH_START = "touchstart",
    TOUCH_MOVE = "touchmove",
    TOUCH_END = "touchend",
    KEY_DOWN = "keydown",
    KEY_PRESS = "keypress",
    KEY_UP = "keyup",
    FOCUS = "focus"
}